+++ project/apps/api/src/app/dto/create-post.dto.ts
@@ -1,4 +1,4 @@
 import { CreatePostDto } from '@project/dto';
-import { OmitType } from '@nestjs/mapped-types';
+import { OmitType } from '@nestjs/swagger';
 
 export class CreatePostApiDto extends OmitType(CreatePostDto, ['author']) {}
