+++ project/libs/shared/app/rdo/README.md
@@ -1,0 +1,7 @@
+# rdo
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test rdo` to execute the unit tests via [Jest](https://jestjs.io).
