+++ project/libs/shared/app/rdo/src/index.ts
@@ -1,0 +1,5 @@
+export * from './lib/file-vault/uploaded-file.rdo';
+
+export * from './lib/users/user.rdo';
+
+export * from './lib/blog/post.rdo';
