+++ project/tsconfig.base.json
@@ -15,13 +15,14 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
+      "@project/config/file-vault": [
+        "libs/shared/config/file-vault/src/index.ts"
+      ],
       "@project/config/notify": ["libs/shared/config/notify/src/index.ts"],
       "@project/config/users": ["libs/shared/config/users/src/index.ts"],
       "@project/dto": ["libs/shared/app/dto/src/index.ts"],
       "@project/libs/shared/app/types": ["libs/shared/app/types/src/index.ts"],
-      "@project/config/file-vault": [
-        "libs/shared/config/file-vault/src/index.ts"
-      ],
+      "@project/rdo": ["libs/shared/app/rdo/src/index.ts"],
       "@project/shared/core": ["libs/shared/core/src/index.ts"],
       "@project/shared/helpers": ["libs/shared/helpers/src/index.ts"]
     },
